requirements.txt
cassandra-driver==3.13.0
celery==3.1.25
Django==2.0.1
django-cassandra-engine==1.4.0
django-celery==3.2.1
django-cors-headers==2.1.0
django-rest-swagger==2.1.2
djangorestframework==3.7.7
kombu==3.0.37
Pillow==4.2.1
psycopg2==2.7.3.1
python-dateutil==2.7.2
redis==2.10.6
requests==2.18.4
